#### Working Build ####
Fixed and tested arrayed assignments.

This example should hopefully clarify what certain assignments mean.

// Begin Example
device res_pack is
 pin[1:0] a = {1,2};
 pin[1:0] b = {3,4};
end device;

net[1:0] myNet;
...
inst(0:1) res_array of res_pack is
 a = myNet & myNet;	// 'a' has 4 bits to assign, i.e. 2 sets of 2.
 this(0).b = myNet;	// 'this(0).b' has 2 bits to assign, i.e 1 set of 2.
 this(1).b = open;  // 'this(1).b' has 2 bits to assign, both to 'open.'
end inst;
// End Example

We are thinking of using a keyword 'each' to handle the other case.

For example:
// Begin Example
// Using the same declarations as the previous example
inst(0:1) res_array of res_pack is
 each a = myNet;	// the user wants myNet to be applied to both instances on the 'a' pins
 this(0).b = myNet;	// 'this(0).b' has 2 bits to assign, i.e 1 set of 2.
 this(1).b = open;  // 'this(1).b' has 2 bits to assign, both to 'open.'
end inst;
// End Example

This is not implemented as of yet.