Board is close to finished.  All parts are put down and connected I think. I shuffled all the LVDS I/O to improve routing.
The PHDL compiler is working very well with the PADS eco flow. Run the compiler then there are three clicks to pull in the changes. More than ever I am convinced that text design entry is better than proprietary, binary schematic tools.

We need to set the differential pairs and create the plane cuts for the special supplies.  Special vias will be required under the AD9739 DAC to escape the BGA pattern. I also want to think about using a copper pour to shield the fast analog sections.

Take a look a the routed design in fmc_dac_blz.pcb!