Hierarchy is working.  Attributes can be modified anywhere in the hierarchy by reaching down with subAttributeAssignment statements.  We decided to clear all attribute "overwritten" flags from all instances inside of a SubInstance.  This dramatically reduces the amount of compiler warnings, especially if lower levels of hierarchy have overwritten attributes specific to that level of hierarchy only.
The JSAP jar was added to the repository (GPL) to assist with command-line parsing and usage generation.  The previous command-line parser was extremely rudimentary and was not as flexible when adding more functionality to the translator.  This JAR will be repackaged into the overall phdl jar upon deployment, likely later tomorrow, along with a reference design (fmc_module with hierarchy) to help everyone get started.