fixed bug in standalone compiler that manifested itself with the upgrade to Xtext 2.3.  Luckily, the developers have provided a copy-and-paste solution until their next maintenance release.  Link: https://bugs.eclipse.org/bugs/show_bug.cgi?id=382555
Also, I have attempted to start on an eclipse product deployment.  Tutorial I found here: http://kthoms.wordpress.com/tag/eclipse-product/.  Although, Step 4 is the most difficult because there is no way to know for sure if all the dependencies are met until I run the application.  This basically reduces to trial and error.  Hopefully they will make this part of deployment easier in the future.