Clarify how to install Vim support on Microsoft Windows.
Different ways of installing Vim on Windows result in different ideas
that Vim may have about where the home directory is, and this affects
where it looks for plugins. Because of that, we direct the user to find
out where $HOME is via Vim itself instead of just guessing where the
path might be.