Add copies of presentations in OpenDocument format.
These existing presentations were in only a proprietary Microsoft
format. Because these presentations are valuable content for the PHDL
repository, I've added a copy of them converted into an open format.

I've left the original files untouched for now, but they probably ought
to be removed and the OpenDocument files used as source material from
now on.