Add GPLv3 COPYING file to top of tree.
Having a file called COPYING with the text of the license is a
more-or-less standard way to provide the license of an open-source work.
There is another HTML-ized copy of the license down deeper in one part
of the repository, but it's hard to find.

It's also important to provide a copy of the applicable license as it is
a requirement of most open source licenses to distribute the license
text along with the program itself.