Removing class and jar files after verifying they can be rebuilt.
There were several compiled class and jar files in the phdl_utils area.
We don't want to store this compiled code in the repository, but before
removing I made sure that we can easily regenerate it from source. This
also serves to double check that the java files all compile.

There is still the issue of 3rd party jar files in this area, but we can
address that separately.