Converted from online Google Doc with permission.
This is some information that Joshua Mangelson wrote up about the
differences between the PHDL 2.0 and PHDL 2.1 grammars. This information
could help us in writing up a language specification document.

Because we want to be as transparent as possible and make it easy to
version control, we want this information to be in our VCS repository.
This is in the standard OpenDocument format, but when we do our actual
spec, we will want this in a better format for version control and
collaboration, such as asciidoc <http://www.methods.co.nz/asciidoc/>.